+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
  * `PackageManager::getPackages()` returns packages with any state now by default
  * `DiscoveryManager::getBindingTypes()` returns types with any state now by default
  * `DiscoveryManager::getBindings()` returns bindings with any state now by default
+ * moved `CompositeKeyStore` to `Puli\RepositoryManager\Util`
 
 * 1.0.0-beta (2015-01-12)